PyProjectConfig does not rely on pydantic anymore
